requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
black==21.10b0
click==8.0.3
Django==3.2.9
djangorestframework==3.12.4
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
